feat: add completion incremental optimazition testcases

Created-by: Cangjie-hljyaozhen
Commit-by: chenminglei
Merged-by: CangjiePL
Description: ## 变更内容（必填）

同步lsp支持增量补全场景优化到main分支，当内存中有可复用的Parse CI，则复用CI，根据上次编译后文件状态变更，增量编译，提高增量补全场景性能

## 变更类型（必填）

请描述本次Pull Request变更类型（原因），**请在保存后点击复选框或在编辑时将对应项前面的 `[ ]` 改为 `[X]` 。**

- [x] 新增需求
- [ ] 问题修复
- [ ] 构建过程或辅助工具变动
- [ ] 文档更新

## 变更内容自检（必填）

**请勿修改或删除以下选项内容，请在保存后点击复选框或在编辑时将对应项前面的 `[ ]` 改为 `[X]` 。**

### 编译本地自验证结果：

- [x] 仓颉工具编译通过
- [ ] 不涉及，辅助工具变动或者文档更新可以直接选择此项

### 测试用例本地自验证结果：

- [x] 通过，请在下方提供截图证明
- [ ] 不涉及，请在下方说明情况
![image.png](https://raw.gitcode.com/user-images/assets/7239752/81f6065b-7fc8-4f9e-9f9d-dddf517ba415/image.png 'image.png')
## 关联的issue（必填）

https://gitcode.com/Cangjie/cangjie_compiler/issues/503

See merge request: Cangjie/cangjie_tools!636

+++ cangjie-language-server/src/languageserver/ArkServer.cpp
@@ -908,8 +908,11 @@ void ArkServer::ChangeWatchedFiles(const std::string &file, FileChangeType type,
                 Logger::Instance().LogMessage(MessageType::MSG_INFO, "recieve file change, but contens are same.");
                 return;
             }
+            auto pkgName = CompilerCangjieProject::GetInstance()->GetFullPkgName(file);
             int64_t version = docMgr->AddDoc(file, 0, contents);
             AddDoc(file, contents, version, ark::NeedDiagnostics::YES, true);
+            CompilerCangjieProject::GetInstance()->
+                UpdateFileStatusInCI(pkgName, file, CompilerInstance::SrcCodeChangeState::CHANGED);
             return;
         }
         if (type == FileChangeType::CREATED) {
@@ -925,7 +928,9 @@ void ArkServer::ChangeWatchedFiles(const std::string &file, FileChangeType type,
         if (type == FileChangeType::DELETED) {
             Logger::Instance().LogMessage(MessageType::MSG_INFO, "delete the file:  " + file);
             CompilerCangjieProject::GetInstance()->IncrementForFileDelete(file);
-            CompilerCangjieProject::GetInstance()->GetBgIndexDB()->DeleteFiles({file});
+            if (CompilerCangjieProject::GetUseDB()) {
+                CompilerCangjieProject::GetInstance()->GetBgIndexDB()->DeleteFiles({file});
+            }
             this->callback->RemoveDocByFile(input.inputs.fileName);
         }
         if (!FileUtil::FileExist(input.onEditFile)) {
